Split analysis.js: pure DSP out, rendering stays
analysis.js mixed the signal maths with the canvas drawing in one 1202-line
file. The 12 pure functions move to src/core/dsp/ — levels (dbfs, clamp,
dbToBar, goertzel), fft (radix-2, Hann, frame accumulation) and spectrum
(Welch averaging, peak finding, interpretation). None of them touched the DOM,
so the split is clean; analysis.js drops to 868 lines and now only draws.

spectrogramColor goes to src/ui/canvas/ instead: it is pure, but a colour ramp
is a presentation choice, not signal processing.

Also lands src/core/constants.ts, because computeSpectrum needs SAMPLE_RATE
and FFT_SIZE and reaching back to a legacy global from core/ would defeat the
point. The four consts leave app.js in the same commit for the usual
shadowing reason.

Corrects the Spectrum type added two commits ago: computeSpectrum returns
Float32Array, not Float64Array as declared. Harmless at runtime, wrong as
documentation — and it would have propagated into every future consumer.

Smoke grows to 45 with a deterministic DSP check that needs no microphone: a
synthetic 1 kHz sine at 20000/32768 full scale comes back as a peak at
1000.0 Hz and -4.3 dBFS, the exact theoretical amplitude. That pins the
windowing and one-sided calibration, which no end-to-end assertion could.

+++ src/core/state.ts
@@ -101,8 +101,8 @@ export interface Recording {
 
 /** Result of an averaged FFT over a selected region. */
 export interface Spectrum {
-  frequencies: Float64Array;
-  magnitudes: Float64Array;
+  frequencies: Float32Array;
+  magnitudes: Float32Array;
   fftSize: number;
   sampleRate: number;
   averagedFrames: number;
